+++ src/main/scala/org/november/learning/akka/RareBooksProtocol.scala
@@ -76,4 +76,81 @@ object RareBooksProtocol {
     def dateInMillis: Long
   }
 
+  /**
+    * Find book by isbn message. call to require function implements business validation rule that ISBN
+    * must by provided/required if the user is searching by ISBN
+    *
+    * @param isbn         isbn to search for
+    * @param dateInMillis date message was created
+    */
+  final case class FindBookByIsbn(isbn: String, dateInMillis: Long = System.currentTimeMillis()) extends Msg {
+    require(isbn.nonEmpty, "Isbn required.")
+  }
+
+  /**
+    * Find book by author message.
+    *
+    * @param author       author to search for
+    * @param dateInMillis date message was created
+    */
+  final case class FindBookByAuthor(author: String, dateInMillis: Long = System.currentTimeMillis()) extends Msg {
+    require(author.nonEmpty, "Author required.")
+  }
+
+  /**
+    * List of book cards found message.
+    *
+    * @param books        list of book cards
+    * @param dateInMillis date message was created
+    */
+  final case class BookFound(books: List[BookCard], dateInMillis: Long = System.currentTimeMillis()) extends Msg {
+    require(books.nonEmpty, "Book(s) required.")
+  }
+
+
+  /**
+    * Find book by title message.
+    *
+    * @param title        title to search for
+    * @param dateInMillis date message was created
+    */
+  final case class FindBookByTitle(title: String, dateInMillis: Long = System.currentTimeMillis()) extends Msg {
+    require(title.nonEmpty, "Title required.")
+  }
+
+  /**
+    * Find book by topic.
+    *
+    * @param topic        set of topics to search for
+    * @param dateInMillis date message was created
+    */
+  final case class FindBookByTopic(topic: Set[Topic], dateInMillis: Long = System.currentTimeMillis()) extends Msg {
+    require(topic.nonEmpty, "Topic required.")
+  }
+
+  /**
+    * Book was not found message.
+    *
+    * @param reason       reason book was not found
+    * @param dateInMillis date message was created
+    */
+  final case class BookNotFound(reason: String, dateInMillis: Long = System.currentTimeMillis()) extends Msg {
+    require(reason.nonEmpty, "Reason is required.")
+  }
+
+  /**
+    * Complain message when book not found.
+    *
+    * @param dateInMillis date message was created
+    */
+  final case class Complain(dateInMillis: Long = System.currentTimeMillis()) extends Msg
+
+  /**
+    * Credit message
+    *
+    * @param dateInMillis date message was created
+    */
+  final case class Credit(dateInMillis: Long = System.currentTimeMillis()) extends Msg
+
+
 }